+++ src/core/ambient.ts
@@ -2,11 +2,11 @@ import fs from "node:fs"
 import path from "node:path"
 import process from "node:process"
 
-const templatesPath = path.join(process.env.TEMPLATES_PATH ?? "./templates")
+const DEFAULT_TEMPLATE_PATH = path.join(process.env.TEMPLATE_PATH ?? "./")
 
 function configureAmbient() {
-  if (!fs.existsSync(templatesPath))
-    fs.mkdirSync(templatesPath, { recursive: true })
+  if (!fs.existsSync(DEFAULT_TEMPLATE_PATH))
+    fs.mkdirSync(DEFAULT_TEMPLATE_PATH, { recursive: true })
 }
 
-export { configureAmbient, templatesPath }
+export { configureAmbient, DEFAULT_TEMPLATE_PATH }