+++ src/core/read.ts
@@ -1,131 +1,163 @@
 import fs from "node:fs"
 import path from "node:path"
 import { basename } from "discord.js"
+import { v4 as uuidv4 } from "uuid"
 import yaml from "yaml"
-import type { ChannelMeta, FsRes, MantoCategory, MantoChannel, MantoConfig, MantoGuild, MantoRole, RoleMeta } from "./types"
-import { MetaType } from "./types.d"
+import type { MantoCategory, MantoChannel, MantoConfig, MantoGuild, MantoRole, MantoServer } from "./types"
+import { existsYaml, getCategoryConfigData, insertMantoId, parseChannelFileName } from "./utils"
 
-function readRes<M extends ChannelMeta | RoleMeta>(filePath: string, lastMeta: M) {
-  const base = path.basename(filePath)
-  const isMeta = base.startsWith("_")
-  const ext = path.extname(base)
-
-  if (ext !== ".yml" && ext !== ".yaml")
-    return null
-
-  const data: Readonly<any> = yaml.parse(fs.readFileSync(filePath, "utf-8"))
-
-  if (base === (`${MetaType.CATEGORY}.yml`) || base === (`${MetaType.CATEGORY}.yaml`))
-    lastMeta.category_name = data.name
-
-  if (base === (`${MetaType.PERMS}.yml`) || base === (`${MetaType.PERMS}.yaml`))
-    lastMeta.perms = { ...lastMeta.perms, ...data }
-
-  return { isMeta, data }
+export function listLiteralFolders(templatePath: string) {
+  return fs.readdirSync(templatePath).filter((file) => {
+    const filePath = path.join(templatePath, file)
+    const stat = fs.statSync(filePath)
+    return stat.isDirectory() && !file.startsWith("_") && !file.startsWith(".")
+  })
 }
 
-export type ReadFilesCb = (res: FsRes) => void
+function readServer(mainFolderPath: string): MantoServer {
+  const mainFolderName = path.basename(mainFolderPath)
 
-function readReses<M extends object>(
-  dirPath: string,
-  cb?: ReadFilesCb,
-  lastMeta: M = {} as M,
-) {
-  const files = fs.readdirSync(dirPath).sort((_, b) => {
-    if (b === "_category.yml" || b === "_category.yaml")
-      return 2
-    if (b.startsWith("_"))
-      return 1
-    return -1
+  let base = existsYaml(mainFolderPath, "_server")
+
+  let data: MantoServer | null = null
+
+  if (!base) {
+    data = {
+      id: uuidv4(),
+    } as MantoServer
+
+    const filePath = path.join(mainFolderPath, "_server.yml")
+    fs.writeFileSync(filePath, yaml.stringify(data))
+    base = "_server.yml"
+  }
+  else {
+    data = yaml.parse(
+      fs.readFileSync(path.join(mainFolderPath, base), "utf-8"),
+    )
+  }
+
+  if (data === null) {
+    const id = uuidv4()
+    const filePath = path.join(mainFolderPath, base)
+    insertMantoId(filePath, id)
+
+    return {
+      id,
+      name: mainFolderName,
+      roles: [],
+      filePath,
+    }
+  }
+  else {
+    const id = uuidv4()
+    const filePath = path.join(mainFolderPath, base)
+    insertMantoId(filePath, id)
+  }
+
+  const roles = readRoles(data)
+
+  delete (data as any).roles
+
+  return {
+    ...data,
+    name: data?.name || mainFolderName,
+    roles,
+    filePath: path.join(mainFolderPath, base),
+  }
+}
+
+function readCategories(dirPath: string) {
+  const categories: MantoCategory[] = []
+  const channels: MantoChannel[] = []
+
+  fs.readdirSync(dirPath).forEach((fileName) => {
+    const filePath = path.join(dirPath, fileName)
+    const stat = fs.statSync(filePath)
+    const isNonCategory = fileName === "_"
+    const isMeta = fileName.startsWith("_")
+
+    if (!stat.isDirectory() || (isMeta && !isNonCategory))
+      return
+
+    const data = getCategoryConfigData(filePath, isNonCategory)
+    const id = isNonCategory ? null : data?.id || uuidv4()
+
+    const category: MantoCategory = {
+      id,
+      name: fileName,
+      filePath,
+      ...data,
+    }
+
+    const categoryChannels = readChannels(category)
+
+    channels.push(...categoryChannels)
+
+    if (!isNonCategory)
+      categories.push(category)
   })
 
-  files.forEach((file) => {
-    const filePath = path.join(dirPath, file)
+  return { categories, channels }
+}
+
+function readChannels(category: MantoCategory) {
+  const channels: MantoChannel[] = []
+
+  fs.readdirSync(category.filePath).forEach((fileName) => {
+    const filePath = path.join(category.filePath, fileName)
+    const isMeta = fileName.startsWith("_")
     const stat = fs.statSync(filePath)
 
-    if (stat.isDirectory()) {
-      readReses(filePath, cb, lastMeta)
+    if (isMeta)
+      return
+
+    if (!stat.isFile()) {
+      console.log("[WARNING] Invalid file in category folder: ", filePath)
       return
     }
 
-    const res = readRes(filePath, lastMeta)
-
-    if (res === null)
-      return
-
-    cb?.({ ...res, filePath, lastMeta: { ...lastMeta } })
-  })
-}
-
-export function readConfig(folderPath: string): MantoConfig {
-  const roles: MantoRole[] = []
-  const channels: MantoChannel[] = []
-  const categories: MantoCategory[] = []
-
-  const readGuild = (folderPath: string): MantoGuild | null => {
-    let base
-
-    if (fs.existsSync(path.join(folderPath, "server.yml")))
-      base = "server.yml"
-    else if (fs.existsSync(path.join(folderPath, "server.yaml")))
-      base = "server.yaml"
-    else base = null
+    const base = existsYaml(category.filePath, basename(filePath, path.extname(filePath)))
 
     if (base === null)
-      return null
+      return
 
-    const filePath = path.join(folderPath, base)
-    const data = yaml.parse(fs.readFileSync(filePath, "utf-8"))
+    const { name, type } = parseChannelFileName(fileName)
 
-    return data
-  }
+    const data = yaml.parse(fs.readFileSync(path.join(category.filePath, base), "utf-8"))
 
-  const readRoles = (dirPath: string) => {
-    readReses<RoleMeta>(dirPath, (res) => {
-      if (res.isMeta)
-        return
-
-      roles.push({ id: res.filePath, ...res.data })
+    channels.push({
+      id: uuidv4(),
+      category: category.id,
+      filePath,
+      ...(data || {}),
+      name: data?.name || name,
+      type,
     })
-  }
+  })
 
-  const readChannels = (dirPath: string) => {
-    const dirChannels: any[] = []
+  return channels
+}
 
-    readReses<ChannelMeta>(
-      dirPath,
-      (res) => {
-        if (!res.isMeta) {
-          dirChannels.push(res)
+function readRoles(data: Record<string, any>) {
+  const roles: MantoRole[] = []
 
-          return
-        }
+  if (data?.roles === undefined)
+    return roles
 
-        const base = basename(res.filePath)
+  data?.roles?.forEach((role: Record<string, any>, index: number) => {
+    roles.push({
+      ...role,
+      id: role.id || uuidv4(),
+      index,
+    })
+  })
 
-        if (base === (`${MetaType.CATEGORY}.yml`) || base === (`${MetaType.CATEGORY}.yaml`)) {
-          categories.push({
-            id: res.filePath,
-            ...res.data,
-          })
-        }
-      },
-    )
+  return roles
+}
 
-    channels.push(...dirChannels.map((channel) => {
-      return {
-        id: channel.filePath,
-        category: channel.lastMeta.category_name,
-        permissions: channel.lastMeta.perms,
-        ...channel.data,
-      }
-    }))
-  }
+export function readConfig(mainFolderPath: string): MantoConfig {
+  const { roles, ...guild } = readServer(mainFolderPath)
+  const { categories, channels } = readCategories(mainFolderPath)
 
-  const guild = readGuild(folderPath)
-
-  readChannels(path.join(folderPath, "channels"))
-  readRoles(path.join(folderPath, "roles"))
-
-  return { guild, roles, channels, categories }
+  return { guild: guild as MantoGuild, roles, channels, categories }
 }