+++ CHANGELOG.md
@@ -5,3 +5,9 @@ description: All notable changes to this project will be documented in this file
 ---
 
 ## [Unreleased]
+
+### Added
+
+- Added support for channels and categories config
+- Added support for roles config
+- Added support for server config